requirements.txt
requests>=2.31.0,<3.0.0
rich>=13.0.0,<14.0.0
python-dotenv>=1.0.0,<2.0.0
pytz>=2023.3,<2024.0
psutil>=5.9.0,<6.0.0
fastapi>=0.110.0,<1.0.0
uvicorn>=0.27.0,<1.0.0
pytest>=7.0.0,<9.0.0